consolidate region and buffer extraction
