+++ Microsoft.DotNet.Try.Project/SourceTextExtensions.cs
@@ -2,23 +2,43 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.Formatting;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.DotNet.Try.Protocol;
-using Workspace = Microsoft.DotNet.Try.Protocol.Workspace;
+using Buffer = Microsoft.DotNet.Try.Protocol.Buffer;
+using InvalidOperationException = System.InvalidOperationException;
+using Path = System.IO.Path;
 
 namespace Microsoft.DotNet.Try.Project
 {
 
     public static class SourceTextExtensions
     {
-        public static IEnumerable<(string id, string content)> ExtractRegions(this SourceText code, string fileName)
+        private const string FSharpRegionStart = "//#region";
+        private const string FSharpRegionEnd = "//#endregion";
+
+        public static IEnumerable<(BufferId bufferId, TextSpan span, TextSpan outerSpan)> ExtractRegions(this SourceText code, string fileName)
         {
-            List<(SyntaxTrivia startRegion, SyntaxTrivia endRegion, string label)> FindRegions(SyntaxNode syntaxNode)
+            var extension = Path.GetExtension(fileName);
+            switch (extension)
+            {
+                case ".cs":
+                case ".csx":
+                    return ExtractRegionsCSharp(code, fileName);
+                case ".fs":
+                case ".fsx":
+                    return ExtractRegionsFSharp(code, fileName);
+                default:
+                    throw new InvalidOperationException($"Unsupported file extension '{extension}'");
+            }
+        }
+
+        private static IEnumerable<(BufferId bufferId, TextSpan span, TextSpan outerSpan)> ExtractRegionsCSharp(SourceText code, string fileName)
+        {
+            IEnumerable<(SyntaxTrivia startRegion, SyntaxTrivia endRegion, BufferId bufferId)> FindRegions(SyntaxNode syntaxNode)
             {
                 var nodesWithRegionDirectives =
                     from node in syntaxNode.DescendantNodesAndTokens()
@@ -28,49 +48,77 @@ namespace Microsoft.DotNet.Try.Project
                           leadingTrivia.Kind() == SyntaxKind.EndRegionDirectiveTrivia
                     select node;
 
-                var regions = new List<(SyntaxTrivia startRegion, SyntaxTrivia endRegion, string label)>();
                 var stack = new Stack<SyntaxTrivia>();
                 var processedSpans = new HashSet<TextSpan>();
+
                 foreach (var nodeWithRegionDirective in nodesWithRegionDirectives)
                 {
                     var triviaList = nodeWithRegionDirective.GetLeadingTrivia();
 
                     foreach (var currentTrivia in triviaList)
                     {
-                        if (!processedSpans.Add(currentTrivia.FullSpan)) continue;
-
-                        if (currentTrivia.Kind() == SyntaxKind.RegionDirectiveTrivia)
+                        if (processedSpans.Add(currentTrivia.FullSpan))
                         {
-                            stack.Push(currentTrivia);
-                        }
-                        else if (currentTrivia.Kind() == SyntaxKind.EndRegionDirectiveTrivia)
-                        {
-                            var start = stack.Pop();
-                            var regionName = start.ToFullString().Replace("#region", string.Empty).Trim();
-                            var regionId = $"{fileName}@{regionName}";
-                            regions.Add(
-                                (start, currentTrivia, regionId));
+                            if (currentTrivia.Kind() == SyntaxKind.RegionDirectiveTrivia)
+                            {
+                                stack.Push(currentTrivia);
+                            }
+                            else if (currentTrivia.Kind() == SyntaxKind.EndRegionDirectiveTrivia && stack.Count > 0)
+                            {
+                                var start = stack.Pop();
+                                var regionName = start.ToFullString().Replace("#region", string.Empty).Trim();
+                                yield return (start, currentTrivia, new BufferId(fileName, regionName));
+                            }
                         }
                     }
                 }
-
-                return regions;
             }
 
             var sourceCodeText = code.ToString();
             var root = CSharpSyntaxTree.ParseText(sourceCodeText).GetRoot();
-            var extractedRegions = new List<(string regionId, string content)>();
+
             foreach (var (startRegion, endRegion, label) in FindRegions(root))
             {
-                var start = startRegion.GetLocation().SourceSpan.End;
-                var length = endRegion.GetLocation().SourceSpan.Start -
+                var innerStart = startRegion.GetLocation().SourceSpan.End;
+
+                var innerLength = endRegion.GetLocation().SourceSpan.Start -
                              startRegion.GetLocation().SourceSpan.End;
-                var loc = new TextSpan(start, length);
 
-                var content = code.ToString(loc);
+                var innerLoc = new TextSpan(innerStart, innerLength);
 
-                content = FormatSourceCode(content);
-                extractedRegions.Add((label, content));
+                var outerStart = startRegion.GetLocation().SourceSpan.Start;
+                var outerLength = endRegion.GetLocation().SourceSpan.End -
+                                  startRegion.GetLocation().SourceSpan.Start;
+                var outerLoc = new TextSpan(outerStart, outerLength);
+
+                yield return (label, innerLoc, outerLoc);
+            }
+        }
+
+        private static IEnumerable<(BufferId bufferId, TextSpan span, TextSpan outerSpan)> ExtractRegionsFSharp(SourceText code, string fileName)
+        {
+            var extractedRegions = new List<(BufferId, TextSpan, TextSpan)>();
+            var text = code.ToString();
+            int regionTagStartIndex = text.IndexOf(FSharpRegionStart);
+            while (regionTagStartIndex >= 0)
+            {
+                var regionLabelEndIndex = text.IndexOf('\n', regionTagStartIndex);
+                var regionLabelStartIndex = regionTagStartIndex + FSharpRegionStart.Length;
+                var regionLabel = text.Substring(regionLabelStartIndex, regionLabelEndIndex - regionLabelStartIndex).Trim();
+                var regionTagEndIndex = text.IndexOf(FSharpRegionEnd, regionTagStartIndex);
+                if (regionTagEndIndex >= 0)
+                {
+                    var regionEndTagLastIndex = regionTagEndIndex + FSharpRegionEnd.Length;
+                    var contentSpan = new TextSpan(regionLabelEndIndex, regionTagEndIndex - regionLabelEndIndex);
+                    var regionSpan = new TextSpan(regionTagStartIndex, regionEndTagLastIndex - regionTagStartIndex);
+                    extractedRegions.Add((new BufferId(fileName, regionLabel), contentSpan, regionSpan));
+
+                    regionTagStartIndex = text.IndexOf(FSharpRegionStart, regionTagEndIndex);
+                }
+                else
+                {
+                    break;
+                }
             }
 
             return extractedRegions;
@@ -78,70 +126,34 @@ namespace Microsoft.DotNet.Try.Project
 
         public static IEnumerable<Buffer> ExtractBuffers(this SourceText code, string fileName)
         {
-            List<(SyntaxTrivia startRegion, SyntaxTrivia endRegion, string label)> FindRegions(SyntaxNode syntaxNode)
+            var extractedBuffers = new List<Buffer>();
+            foreach ((var bufferId, var contentSpan, var regionSpan) in ExtractRegions(code, fileName))
             {
-                var nodesWithRegionDirectives =
-                    from node in syntaxNode.DescendantNodesAndTokens()
-                    where node.HasLeadingTrivia
-                    from leadingTrivia in node.GetLeadingTrivia()
-                    where leadingTrivia.Kind() == SyntaxKind.RegionDirectiveTrivia ||
-                          leadingTrivia.Kind() == SyntaxKind.EndRegionDirectiveTrivia
-                    select node;
-
-                var regions = new List<(SyntaxTrivia startRegion, SyntaxTrivia endRegion, string label)>();
-                var stack = new Stack<SyntaxTrivia>();
-                var processedSpans = new HashSet<TextSpan>();
-                foreach (var nodeWithRegionDirective in nodesWithRegionDirectives)
-                {
-                    var triviaList = nodeWithRegionDirective.GetLeadingTrivia();
-
-                    foreach (var currentTrivia in triviaList)
-                    {
-                        if (!processedSpans.Add(currentTrivia.FullSpan)) continue;
-
-                        if (currentTrivia.Kind() == SyntaxKind.RegionDirectiveTrivia)
-                        {
-                            stack.Push(currentTrivia);
-                        }
-                        else if (currentTrivia.Kind() == SyntaxKind.EndRegionDirectiveTrivia)
-                        {
-                            var start = stack.Pop();
-                            var regionName = start.ToFullString().Replace("#region", string.Empty).Trim();
-                            var regionId = $"{fileName}@{regionName}";
-                            regions.Add(
-                                (start, currentTrivia, regionId));
-                        }
-                    }
-                }
-
-                return regions;
+                var content = code.ToString(contentSpan);
+                content = content.FormatSourceCode(fileName);
+                extractedBuffers.Add(new Buffer(bufferId, content));
             }
 
-            if (Path.GetExtension(fileName) == ".fs")
-            {
-                return FSharpMethods.ExtractBuffers(code, fileName);
-            }
-
-            var sourceCodeText = code.ToString();
-            var root = CSharpSyntaxTree.ParseText(sourceCodeText).GetRoot();
-            var extractedRegions = new List<Buffer>();
-            foreach (var (startRegion, endRegion, label) in FindRegions(root))
-            {
-                var start = startRegion.GetLocation().SourceSpan.End;
-                var length = endRegion.GetLocation().SourceSpan.Start -
-                             startRegion.GetLocation().SourceSpan.End;
-                var loc = new TextSpan(start, length);
-
-                var content = code.ToString(loc);
-
-                content = FormatSourceCode(content);
-                extractedRegions.Add(new Buffer(label, content));
-            }
-
-            return extractedRegions;
+            return extractedBuffers;
         }
 
-        private static string FormatSourceCode(string sourceCode)
+        public static string FormatSourceCode(this string sourceCode, string fileName)
+        {
+            var extension = Path.GetExtension(fileName);
+            switch (extension)
+            {
+                case ".cs":
+                case ".csx":
+                    return FormatSourceCodeCSharp(sourceCode);
+                case ".fs":
+                case ".fsx":
+                    return sourceCode;
+                default:
+                    throw new InvalidOperationException($"Unsupported file extension '{extension}'");
+            }
+        }
+
+        private static string FormatSourceCodeCSharp(string sourceCode)
         {
             var tree = CSharpSyntaxTree.ParseText(sourceCode.Trim(), new CSharpParseOptions(kind: SourceCodeKind.Script));
             var cw = new AdhocWorkspace();
